+++ skrub/_data_ops/_skrub_namespace.py
@@ -1,3 +1,4 @@
+import functools
 import pickle
 import re
 import sys
@@ -20,6 +21,7 @@ from ._data_ops import (
     Var,
     check_data_op,
     check_name,
+    checked_data_op_constructor,
     deferred,
 )
 from ._estimator import (
@@ -92,6 +94,29 @@ def _check_grid_search_possible(data_op):
             )
 
 
+def _check_before(f):
+    """
+    Decorator to perform validation of a DataOp before calling a function.
+
+    Usually some checks such as no duplicate names are performed whenever a
+    DataOp is created. However to reduce overhead, those checks can be disabled
+    when the DataOp is created. But we always perform validation before
+    actually using the DataOp. So all functions that evaluate the DataOp or
+    transform it into some other type such as .skb.eval(), .skb.make_learner()
+    etc. must be decorated with _check_before.
+
+    Note that once a DataOp has been checked, that is stored in an attribute so
+    redundant checks are avoided.
+    """
+
+    @functools.wraps(f)
+    def _checked(self, *args, **kwargs):
+        check_data_op(self._data_op)
+        return f(self, *args, **kwargs)
+
+    return _checked
+
+
 def _is_optuna_trial(obj):
     try:
         optuna = sys.modules["optuna"]
@@ -138,7 +163,7 @@ class SkrubNamespace:
         )
         return data_op
 
-    @check_data_op
+    @checked_data_op_constructor
     def apply(
         self,
         estimator,
@@ -489,7 +514,7 @@ class SkrubNamespace:
         """
         return deferred(func)(self._data_op, *args, **kwargs)
 
-    @check_data_op
+    @checked_data_op_constructor
     def if_else(self, value_if_true, value_if_false):
         """Create a conditional DataOp.
 
@@ -552,7 +577,7 @@ class SkrubNamespace:
         """
         return DataOp(IfElse(self._data_op, value_if_true, value_if_false))
 
-    @check_data_op
+    @checked_data_op_constructor
     def match(self, targets, default=NULL):
         """Select based on the value of a DataOp.
 
@@ -615,7 +640,7 @@ class SkrubNamespace:
         """
         return DataOp(Match(self._data_op, targets, default))
 
-    @check_data_op
+    @checked_data_op_constructor
     def select(self, cols):
         """Select a subset of columns.
 
@@ -668,7 +693,7 @@ class SkrubNamespace:
         """
         return self._apply(SelectCols(cols), how="no_wrap")
 
-    @check_data_op
+    @checked_data_op_constructor
     def drop(self, cols):
         """Drop some columns.
 
@@ -721,7 +746,7 @@ class SkrubNamespace:
         """
         return self._apply(DropCols(cols), how="no_wrap")
 
-    @check_data_op
+    @checked_data_op_constructor
     def concat(self, others, axis=0):
         """Concatenate dataframes vertically or horizontally.
 
@@ -781,7 +806,7 @@ class SkrubNamespace:
         """
         return DataOp(Concat(self._data_op, others, axis=axis))
 
-    @check_data_op
+    @checked_data_op_constructor
     def subsample(self, n=1000, *, how="head"):
         """Configure subsampling of a dataframe or numpy array.
 
@@ -945,6 +970,7 @@ class SkrubNamespace:
         """  # noqa : E501
         return DataOp(SubsamplePreviews(self._data_op, n=n, how=how))
 
+    @_check_before
     def clone(self, drop_values=True):
         """Get an independent clone of the DataOp.
 
@@ -1002,6 +1028,7 @@ class SkrubNamespace:
 
         return clone(self._data_op, drop_preview_data=drop_values)
 
+    @_check_before
     def eval(self, environment=None, *, keep_subsampling=False):
         """Evaluate the DataOp.
 
@@ -1073,6 +1100,7 @@ class SkrubNamespace:
             self._data_op, mode="fit_transform", environment=environment, clear=True
         )
 
+    @_check_before
     def preview(self):
         """Get the value computed for previews (shown when printing the DataOp).
 
@@ -1127,7 +1155,7 @@ class SkrubNamespace:
         """
         return evaluate(self._data_op, mode="preview", environment=None, clear=False)
 
-    @check_data_op
+    @checked_data_op_constructor
     def freeze_after_fit(self):
         """Freeze the result during learner fitting.
 
@@ -1178,6 +1206,7 @@ class SkrubNamespace:
         """
         return DataOp(FreezeAfterFit(self._data_op))
 
+    @_check_before
     def get_data(self):
         """Collect the values of the variables contained in the DataOp.
 
@@ -1205,6 +1234,7 @@ class SkrubNamespace:
                 data[impl.name] = impl.value
         return data
 
+    @_check_before
     def get_vars(self, all_named_ops=False):
         """
         Get all the variables used in the DataOp.
@@ -1291,6 +1321,7 @@ class SkrubNamespace:
             if isinstance(op._skrub_impl, Var)
         }
 
+    @_check_before
     def draw_graph(self):
         """Get an SVG string representing the computation graph.
 
@@ -1308,6 +1339,7 @@ class SkrubNamespace:
 
         return draw_data_op_graph(self._data_op)
 
+    @_check_before
     def describe_steps(self):
         """Get a text representation of the computation graph.
 
@@ -1356,6 +1388,7 @@ class SkrubNamespace:
 
         return describe_steps(self._data_op)
 
+    @_check_before
     def describe_param_grid(self):
         """Describe the hyper-parameters extracted from choices in the DataOp.
 
@@ -1438,6 +1471,7 @@ class SkrubNamespace:
 
         return describe_param_grid(self._data_op)
 
+    @_check_before
     def describe_defaults(self):
         """Describe the hyper-parameters used by the default learner.
 
@@ -1473,6 +1507,7 @@ class SkrubNamespace:
 
         return describe_params(eval_choices(self._data_op), choice_graph(self._data_op))
 
+    @_check_before
     def full_report(
         self,
         environment=None,
@@ -1600,6 +1635,7 @@ class SkrubNamespace:
             title=title,
         )
 
+    @_check_before
     def make_learner(self, *, fitted=False, keep_subsampling=False, choose="default"):
         """Get a skrub learner for this DataOp.
 
@@ -1772,6 +1808,7 @@ class SkrubNamespace:
             env_with_subsampling(self._data_op, self.get_data(), keep_subsampling)
         )
 
+    @_check_before
     def train_test_split(
         self,
         environment=None,
@@ -1865,6 +1902,7 @@ class SkrubNamespace:
             **split_func_kwargs,
         )
 
+    @_check_before
     def iter_cv_splits(self, environment=None, *, keep_subsampling=False, cv=KFOLD_5):
         """Yield splits of an environment into training and testing environments.
 
@@ -1930,6 +1968,7 @@ class SkrubNamespace:
             self._data_op, environment, keep_subsampling=keep_subsampling, cv=cv
         )
 
+    @_check_before
     def make_grid_search(self, *, fitted=False, keep_subsampling=False, **kwargs):
         """Find the best parameters with grid search.
 
@@ -2041,6 +2080,7 @@ class SkrubNamespace:
             env_with_subsampling(self._data_op, self.get_data(), keep_subsampling)
         )
 
+    @_check_before
     def make_randomized_search(
         self,
         *,
@@ -2327,6 +2367,7 @@ class SkrubNamespace:
             env_with_subsampling(self._data_op, self.get_data(), keep_subsampling)
         )
 
+    @_check_before
     def iter_learners_grid(self):
         """Get learners with different parameter combinations.
 
@@ -2408,6 +2449,7 @@ class SkrubNamespace:
             new.set_params(**params)
             yield new
 
+    @_check_before
     def iter_learners_randomized(self, n_iter, *, random_state=None):
         """Get learners with different parameter combinations.
 
@@ -2479,6 +2521,7 @@ class SkrubNamespace:
             new.set_params(**params)
             yield new
 
+    @_check_before
     def cross_validate(self, environment=None, *, keep_subsampling=False, **kwargs):
         """Cross-validate the DataOp plan.
 
@@ -2543,7 +2586,7 @@ class SkrubNamespace:
             **kwargs,
         )
 
-    @check_data_op
+    @checked_data_op_constructor
     def mark_as_X(self):
         """Mark this DataOp as being the ``X`` table.
 
@@ -2625,7 +2668,7 @@ class SkrubNamespace:
         """Whether this DataOp has been marked with :meth:`.skb.mark_as_X()`."""
         return self._data_op._skrub_impl.is_X
 
-    @check_data_op
+    @checked_data_op_constructor
     def mark_as_y(self):
         """Mark this DataOp as being the ``y`` table.
 
@@ -2701,7 +2744,7 @@ class SkrubNamespace:
         """Whether this DataOp has been marked with :meth:`.skb.mark_as_y()`."""
         return self._data_op._skrub_impl.is_y
 
-    @check_data_op
+    @checked_data_op_constructor
     def set_name(self, name):
         """Give a name to this DataOp.
 
@@ -2827,7 +2870,7 @@ class SkrubNamespace:
         return f"<{self.__class__.__name__}>"
 
     @property
-    @check_data_op
+    @checked_data_op_constructor
     def applied_estimator(self):
         """Retrieve the estimator applied in the previous step, as a DataOp.